Add hooks/handlers for application->client interface.
Applications can now connect to the Mercury Client Endpoint and request
subscriptions, send messages, etc. I also updated the Broker to specify
the type of safety message it is sending, and moved these to the
'Broker_Safety' topic.  I added a 'Broker_Utility' topic for handling
echo requests (for testing) and other such things we may come up with.

Overall the handling of pubsub messages is much cleaner now.